refactored code (read commit description)
The code (namely storage logic) is a lot more readable and maintainable in preparation of the next version.

Looking forward, I'll do my best to make smaller changes and to divide them in more commits, as I've been neglectful to people who carefully read the changes in the code for each update. Sorry for that :)